+++ lib/src/main/java/com/orange/ods/compose/component/controls/OdsSwitch.kt
@@ -1,0 +1,47 @@
+/*
+ *
+ *  Copyright 2021 Orange
+ *
+ *  Use of this source code is governed by an MIT-style
+ *  license that can be found in the LICENSE file or at
+ *  https://opensource.org/licenses/MIT.
+ * /
+ */
+
+package com.orange.ods.compose.component.controls
+
+import androidx.compose.material.Switch
+import androidx.compose.material.ripple.LocalRippleTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.ui.Modifier
+import com.orange.ods.compose.theme.OdsPrimaryRippleTheme
+
+/**
+ * <a href="https://system.design.orange.com/0c1af118d/p/14638a-selection-controls/b/352c00" class="external" target="_blank">ODS switch</a>.
+ *
+ * Switches toggle the state of a single item on or off.
+ *
+ * @param checked whether or not this component is checked
+ * @param onCheckedChange callback to be invoked when Switch is being clicked,
+ * therefore the change of checked state is requested.  If null, then this is passive
+ * and relies entirely on a higher-level component to control the "checked" state.
+ * @param modifier Modifier to be applied to the switch layout
+ * @param enabled whether the component is enabled or grayed out
+ */
+@Composable
+fun OdsSwitch(
+    checked: Boolean,
+    onCheckedChange: ((Boolean) -> Unit)?,
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+) {
+    CompositionLocalProvider(LocalRippleTheme provides OdsPrimaryRippleTheme) {
+        Switch(
+            checked = checked,
+            onCheckedChange = onCheckedChange,
+            modifier = modifier,
+            enabled = enabled,
+        )
+    }
+}